Highlithed current rules in the options page
