+++ .eslintignore
@@ -1,2 +1,3 @@
 asocial/content
 asocial/options/options.js
+asocial/event
