+++ asocial/src/utils.js
@@ -37,5 +37,44 @@ module.exports = {
         array.forEach(df.appendChild, df);
 
         return df;
+    },
+
+    /**
+     * checkRule - checking rule
+     *
+     * @param  {Date} time
+     * @param  {Object} rule
+     * @param  {String} network
+     * @return {Boolean}
+     */
+    checkRule: function checkRule(time, rule, network) {
+        if (network) {
+            var disabledNetworks = Object.keys(rule.sites).filter(network => rule.sites[network]);
+
+            if (disabledNetworks.length > 0 && !rule.sites[network]) {
+                return false;
+            }
+        }
+
+        if (rule.days.length !== 0 && rule.days.indexOf(time.getDay()) === -1) {
+            return false;
+        }
+
+        var startTime = new Date();
+        var endTime = new Date();
+        var isAfterStart = true;
+        var isBeforeEnd = true;
+
+        if (rule.start) {
+            startTime.setHours(...rule.start, 0, 0);
+            isAfterStart = startTime <= time;
+        }
+
+        if (rule.end) {
+            endTime.setHours(...rule.end, 0, 0);
+            isBeforeEnd = time < endTime;
+        }
+
+        return isAfterStart && isBeforeEnd;
     }
 };
